// src/lib/mock-api/types.ts
export type ContentStatus =
  | "draft"
  | "pending"
  | "published"
  | "scheduled"
  | "private"
  | "unlisted"
  | "restricted"
  | "rejected"
  | "archived";
// ...
export type CampaignStatus =
  | "draft"
  | "pending"
  | "active"
  | "paused"
  | "completed"
  | "rejected";